Fix incorrect usage of `unsigned` and `int` in place of NSUInteger/NSInteger's.
On x64, NSInteger is 8 bytes long while `int` is only 4. This causes problems when ex. NSNotFound
is cast to an int and then compared to itself.

There probably are many more places the integer types are used the wrong way, and this is likely to
cause us more headache later.

+++ AppKit/NSAttributedString.m
@@ -182,7 +182,7 @@ NSString * const NSSpellingStateAttributeName=@"NSSpellingStateAttributeName"; /
 }
 
 /* as usual, the documentation says one thing and the system behaves differently, this is the way i think it should work... (dwy 5/11/2003) */
--(unsigned)nextWordFromIndex:(unsigned)location forward:(BOOL)forward {
+-(NSUInteger)nextWordFromIndex:(NSUInteger)location forward:(BOOL)forward {
     NSCharacterSet *alpha = [NSCharacterSet alphanumericCharacterSet];
     NSString *string = [self string];
     int i = location, length = [self length];
