+++ AppKit/NSMenu.subproj/NSMenuView.h
@@ -9,7 +9,7 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
 #import <AppKit/AppKit.h>
 
 @interface NSMenuView : NSView {
-    unsigned _selectedItemIndex;
+    NSUInteger _selectedItemIndex;
     NSMutableArray *_visibleArray;
 }
